+++ test/source/lob_env_test.cpp
@@ -27,13 +27,13 @@ struct LobEnvTestFixture : public testing::Test {
     puut = std::make_unique<lob::Builder>();
     ASSERT_NE(puut, nullptr);
 
-    const float kTestBC = 0.232F;
+    const double kTestBC = 0.232;
     const lob::DragFunctionT kDragFunction = lob::DragFunctionT::kG7;
-    const float kTestDiameter = 0.308F;
-    const float kTestWeight = 155.0F;
+    const double kTestDiameter = 0.308;
+    const double kTestWeight = 155.0;
     const uint16_t kTestMuzzleVelocity = 2800;
-    const float kTestZeroAngle = 3.66F;
-    const float kTestOpticHeight = 1.5F;
+    const double kTestZeroAngle = 3.66;
+    const double kTestOpticHeight = 1.5;
 
     puut->BallisticCoefficientPsi(kTestBC)
         .BCDragFunction(kDragFunction)
@@ -51,19 +51,19 @@ struct LobEnvTestFixture : public testing::Test {
 TEST_F(LobEnvTestFixture, ZeroAngleSearch) {
   ASSERT_NE(puut, nullptr);
   auto input1 = puut->Build();
-  const float kZeroRange = 100.0F;
-  auto input2 = puut->ZeroAngleMOA(std::numeric_limits<float>::quiet_NaN())
+  const double kZeroRange = 100.0;
+  auto input2 = puut->ZeroAngleMOA(std::numeric_limits<double>::quiet_NaN())
                     .ZeroDistanceYds(kZeroRange)
                     .Build();
-  const float kError = 0.01F;
+  const double kError = 0.01;
   EXPECT_NEAR(input1.zero_angle, input2.zero_angle, kError);
 }
 
 TEST_F(LobEnvTestFixture, GetSpeedOfSoundFps) {
   ASSERT_NE(puut, nullptr);
   const auto kInput = puut->Build();
-  const float kExpectedFps = 1116.45F;
-  const float kError = 0.001F;
+  const double kExpectedFps = 1116.45;
+  const double kError = 0.001;
   EXPECT_NEAR(kInput.speed_of_sound, kExpectedFps, kError);
 }
 
@@ -79,18 +79,18 @@ TEST_F(LobEnvTestFixture, SolveAtICAOAtmosphere) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2699, 2505, -0.15F, 0.00F, 0.055F},
-      {300, 2600, 2325, 0.00F, 0.00F, 0.111F},
-      {600, 2409, 1995, -3.64F, 0.00F, 0.231F},
-      {900, 2225, 1703, -13.30F, 0.00F, 0.361F},
-      {1200, 2051, 1446, -29.98F, 0.00F, 0.501F},
-      {1500, 1883, 1220, -54.96F, 0.00F, 0.654F},
-      {1800, 1723, 1021, -89.76F, 0.00F, 0.820F},
-      {2100, 1569, 846, -136.31F, 0.00F, 1.003F},
-      {2400, 1421, 694, -197.03F, 0.00F, 1.204F},
-      {2700, 1280, 564, -275.05F, 0.00F, 1.426F},
-      {3000, 1149, 454, -374.36F, 0.00F, 1.674F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2699, 2505, -0.15, 0.00, 0.055},
+      {300, 2600, 2325, 0.00, 0.00, 0.111},
+      {600, 2409, 1995, -3.64, 0.00, 0.231},
+      {900, 2225, 1703, -13.30, 0.00, 0.361},
+      {1200, 2051, 1446, -29.98, 0.00, 0.501},
+      {1500, 1883, 1220, -54.96, 0.00, 0.654},
+      {1800, 1723, 1021, -89.76, 0.00, 0.820},
+      {2100, 1569, 846, -136.31, 0.00, 1.003},
+      {2400, 1421, 694, -197.03, 0.00, 1.204},
+      {2700, 1280, 564, -275.05, 0.00, 1.426},
+      {3000, 1149, 454, -374.36, 0.00, 1.674}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
@@ -131,18 +131,18 @@ TEST_F(LobEnvTestFixture, SolveWithAltitude4500ft) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2714, 2533, -0.15F, 0.00F, 0.054F},
-      {300, 2630, 2379, 0.02F, 0.00F, 0.111F},
-      {600, 2466, 2091, -3.49F, 0.00F, 0.228F},
-      {900, 2308, 1831, -12.72F, 0.00F, 0.354F},
-      {1200, 2156, 1598, -28.49F, 0.00F, 0.488F},
-      {1500, 2010, 1389, -51.76F, 0.00F, 0.633F},
-      {1800, 1869, 1201, -83.62F, 0.00F, 0.787F},
-      {2100, 1733, 1033, -125.48F, 0.00F, 0.954F},
-      {2400, 1602, 882, -178.94F, 0.00F, 1.134F},
-      {2700, 1475, 748, -245.97F, 0.00F, 1.329F},
-      {3000, 1353, 629, -329.05F, 0.00F, 1.542F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2714, 2533, -0.15, 0.00, 0.054},
+      {300, 2630, 2379, 0.02, 0.00, 0.111},
+      {600, 2466, 2091, -3.49, 0.00, 0.228},
+      {900, 2308, 1831, -12.72, 0.00, 0.354},
+      {1200, 2156, 1598, -28.49, 0.00, 0.488},
+      {1500, 2010, 1389, -51.76, 0.00, 0.633},
+      {1800, 1869, 1201, -83.62, 0.00, 0.787},
+      {2100, 1733, 1033, -125.48, 0.00, 0.954},
+      {2400, 1602, 882, -178.94, 0.00, 1.134},
+      {2700, 1475, 748, -245.97, 0.00, 1.329},
+      {3000, 1353, 629, -329.05, 0.00, 1.542}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
@@ -182,18 +182,18 @@ TEST_F(LobEnvTestFixture, SolveWithTempAndAirPressure) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2720, 2544, -0.15F, 0.00F, 0.054F},
-      {300, 2642, 2400, 0.03F, 0.00F, 0.110F},
-      {600, 2489, 2130, -3.43F, 0.00F, 0.227F},
-      {900, 2341, 1885, -12.50F, 0.00F, 0.351F},
-      {1200, 2199, 1662, -27.94F, 0.00F, 0.484F},
-      {1500, 2061, 1461, -50.56F, 0.00F, 0.625F},
-      {1800, 1929, 1279, -81.36F, 0.00F, 0.775F},
-      {2100, 1800, 1114, -121.56F, 0.00F, 0.936F},
-      {2400, 1675, 965, -172.49F, 0.00F, 1.109F},
-      {2700, 1554, 830, -235.82F, 0.00F, 1.295F},
-      {3000, 1437, 710, -313.59F, 0.00F, 1.496F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2720, 2544, -0.15, 0.00, 0.054},
+      {300, 2642, 2400, 0.03, 0.00, 0.110},
+      {600, 2489, 2130, -3.43, 0.00, 0.227},
+      {900, 2341, 1885, -12.50, 0.00, 0.351},
+      {1200, 2199, 1662, -27.94, 0.00, 0.484},
+      {1500, 2061, 1461, -50.56, 0.00, 0.625},
+      {1800, 1929, 1279, -81.36, 0.00, 0.775},
+      {2100, 1800, 1114, -121.56, 0.00, 0.936},
+      {2400, 1675, 965, -172.49, 0.00, 1.109},
+      {2700, 1554, 830, -235.82, 0.00, 1.295},
+      {3000, 1437, 710, -313.59, 0.00, 1.496}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
@@ -237,18 +237,18 @@ TEST_F(LobEnvTestFixture, SolveWithBarometricPressure) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2716, 2537, -0.15F, 0.00F, 0.054F},
-      {300, 2634, 2385, 0.02F, 0.00F, 0.110F},
-      {600, 2472, 2102, -3.47F, 0.00F, 0.228F},
-      {900, 2317, 1846, -12.66F, 0.00F, 0.353F},
-      {1200, 2168, 1615, -28.33F, 0.00F, 0.487F},
-      {1500, 2024, 1408, -51.41F, 0.00F, 0.630F},
-      {1800, 1885, 1222, -83.00F, 0.00F, 0.784F},
-      {2100, 1752, 1055, -124.38F, 0.00F, 0.949F},
-      {2400, 1622, 905, -177.12F, 0.00F, 1.127F},
-      {2700, 1497, 770, -243.08F, 0.00F, 1.320F},
-      {3000, 1376, 651, -324.63F, 0.00F, 1.529F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2716, 2537, -0.15, 0.00, 0.054},
+      {300, 2634, 2385, 0.02, 0.00, 0.110},
+      {600, 2472, 2102, -3.47, 0.00, 0.228},
+      {900, 2317, 1846, -12.66, 0.00, 0.353},
+      {1200, 2168, 1615, -28.33, 0.00, 0.487},
+      {1500, 2024, 1408, -51.41, 0.00, 0.630},
+      {1800, 1885, 1222, -83.00, 0.00, 0.784},
+      {2100, 1752, 1055, -124.38, 0.00, 0.949},
+      {2400, 1622, 905, -177.12, 0.00, 1.127},
+      {2700, 1497, 770, -243.08, 0.00, 1.320},
+      {3000, 1376, 651, -324.63, 0.00, 1.529}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
@@ -291,18 +291,18 @@ TEST_F(LobEnvTestFixture, SolveWithPressureTempHumidity) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2705, 2516, -0.15F, 0.00F, 0.055F},
-      {300, 2612, 2346, 0.01F, 0.00F, 0.111F},
-      {600, 2431, 2033, -3.58F, 0.00F, 0.230F},
-      {900, 2258, 1753, -13.06F, 0.00F, 0.358F},
-      {1200, 2092, 1505, -29.38F, 0.00F, 0.496F},
-      {1500, 1934, 1285, -53.64F, 0.00F, 0.645F},
-      {1800, 1781, 1090, -87.24F, 0.00F, 0.807F},
-      {2100, 1633, 917, -131.81F, 0.00F, 0.983F},
-      {2400, 1491, 765, -189.49F, 0.00F, 1.175F},
-      {2700, 1355, 632, -262.81F, 0.00F, 1.386F},
-      {3000, 1227, 517, -355.17F, 0.00F, 1.619F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2705, 2516, -0.15, 0.00, 0.055},
+      {300, 2612, 2346, 0.01, 0.00, 0.111},
+      {600, 2431, 2033, -3.58, 0.00, 0.230},
+      {900, 2258, 1753, -13.06, 0.00, 0.358},
+      {1200, 2092, 1505, -29.38, 0.00, 0.496},
+      {1500, 1934, 1285, -53.64, 0.00, 0.645},
+      {1800, 1781, 1090, -87.24, 0.00, 0.807},
+      {2100, 1633, 917, -131.81, 0.00, 0.983},
+      {2400, 1491, 765, -189.49, 0.00, 1.175},
+      {2700, 1355, 632, -262.81, 0.00, 1.386},
+      {3000, 1227, 517, -355.17, 0.00, 1.619}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
@@ -330,11 +330,11 @@ TEST_F(LobEnvTestFixture, SolveWithPressureTempHumidity) {
 
 // NOLINTNEXTLINE(readability-function-cognitive-complexity)
 TEST_F(LobEnvTestFixture, SolveWithWeatherStationData) {
-  const float kAltitudeOfFiringSite = 5'280.0F;
-  const float kAirPressure = 30.0F;
-  const float kAltitudeOfBarometer = 0.0F;
-  const float kTemperature = 65.0F;
-  const float kAltitudeOfThermometer = 3'598.0F;
+  const double kAltitudeOfFiringSite = 5'280.0;
+  const double kAirPressure = 30.0;
+  const double kAltitudeOfBarometer = 0.0;
+  const double kTemperature = 65.0;
+  const double kAltitudeOfThermometer = 3'598.0;
   const auto kInput = puut->AltitudeOfFiringSiteFt(kAltitudeOfFiringSite)
                           .AirPressureInHg(kAirPressure)
                           .AltitudeOfBarometerFt(kAltitudeOfBarometer)
@@ -350,18 +350,18 @@ TEST_F(LobEnvTestFixture, SolveWithWeatherStationData) {
   const std::array<uint32_t, kSolutionLength> kRanges = {
       0, 150, 300, 600, 900, 1200, 1500, 1800, 2100, 2400, 2700, 3000};
   const std::vector<lob::Output> kExpected = {
-      {0, 2800, 2696, -1.50F, 0.00F, 0.000F},
-      {150, 2716, 2537, -0.15F, 0.00F, 0.054F},
-      {300, 2634, 2385, 0.02F, 0.00F, 0.110F},
-      {600, 2472, 2102, -3.47F, 0.00F, 0.228F},
-      {900, 2317, 1846, -12.66F, 0.00F, 0.353F},
-      {1200, 2168, 1615, -28.33F, 0.00F, 0.487F},
-      {1500, 2024, 1408, -51.41F, 0.00F, 0.630F},
-      {1800, 1885, 1222, -83.00F, 0.00F, 0.784F},
-      {2100, 1752, 1055, -124.38F, 0.00F, 0.949F},
-      {2400, 1622, 905, -177.12F, 0.00F, 1.127F},
-      {2700, 1497, 770, -243.08F, 0.00F, 1.320F},
-      {3000, 1376, 651, -324.63F, 0.00F, 1.529F}};
+      {0, 2800, 2696, -1.50, 0.00, 0.000},
+      {150, 2716, 2537, -0.15, 0.00, 0.054},
+      {300, 2634, 2385, 0.02, 0.00, 0.110},
+      {600, 2472, 2102, -3.47, 0.00, 0.228},
+      {900, 2317, 1846, -12.66, 0.00, 0.353},
+      {1200, 2168, 1615, -28.33, 0.00, 0.487},
+      {1500, 2024, 1408, -51.41, 0.00, 0.630},
+      {1800, 1885, 1222, -83.00, 0.00, 0.784},
+      {2100, 1752, 1055, -124.38, 0.00, 0.949},
+      {2400, 1622, 905, -177.12, 0.00, 1.127},
+      {2700, 1497, 770, -243.08, 0.00, 1.320},
+      {3000, 1376, 651, -324.63, 0.00, 1.529}};
 
   std::array<lob::Output, kSolutionLength> solutions = {};
   const lob::Options kOptions = {0, 0, lob::NaN(), kTestStepSize};
